monitor-system/requirements.txt
#kafka==1.3.5
kafka_python==2.0.2
psutil==5.9.4
